Move fps_counter
